// src/schemas/yupSchema.ts
import * as yup from "yup"; 

export const registerSchema = yup.object({
    name: yup.string().required("Name is required"),
    email: yup.string().required("Email is required").email("Enter a valid email"), 
    password: yup.string().required("Password is required").min(6, "Password must be at least 6 characters") 
}).required(); 

export const loginShema = yup.object({
    email: yup.string().required("Email is required").email("Enter a valid email"), 
    password: yup.string().required("Password is required").min(6, "Password must be at least 6 characters") 
}).required();